+++ handlers/delivery-records-api/src/main/scala/com/gu/delivery_records_api/Handler.scala
@@ -1,6 +1,6 @@
 package com.gu.delivery_records_api
 
-import cats.effect.IO
+import cats.effect.{ContextShift, IO}
 import cats.syntax.either._
 import com.gu.http4s.Http4sLambdaHandler
 import sttp.client3.asynchttpclient.cats.AsyncHttpClientCatsBackend
@@ -9,10 +9,10 @@ import scala.concurrent.ExecutionContext
 
 object Handler
     extends Http4sLambdaHandler({
-      implicit val contextShift = IO.contextShift(ExecutionContext.global)
+      implicit val contextShift: ContextShift[IO] = IO.contextShift(ExecutionContext.global)
       AsyncHttpClientCatsBackend[IO]()
         .flatMap { sttpBackend =>
-          DeliveryRecordsApiApp(sttpBackend).value
+          DeliveryRecordsApiApp.buildHttpRoutes(sttpBackend).value
         }
         .unsafeRunSync()
         .valueOr((error: DeliveryRecordsApiError) => throw new RuntimeException(error.toString))
